+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 - **[NEW]** Added `display` prop to `Proximity`
 - **[NEW]** Brand new `Itinerary` component (under `newItinerary` folder)
-[...]
-
+- **[NEW]** Add `BlueGradient` in colors
+- **[NEW]** Add `Brand` in typography
+- **[NEW]** Add `BrandedVoice` component
 - **[FIX]** Fix `Success Screen` gutter for large screens
 - **[FIX]** Remove background from `SEA Pages`
-
+[...]
 
 # v47.5.0 (12/02/2021)